Moved theme to a different file, updated styles to be generated from theme
